Add files via upload
my first object oriented programing in c++ . The code demonstrates a bank management system on console . It was developed by using the classes in c++ and basic 
do-while loop logic . It contains private  data inside class which stores the account number, account holder, and the initial balance of that account following information are to be entered in the main function. further this balance goes through different operations depending on the customer use.